권한 없이도 reset.php에 접속하면 초기화 가능한 문제 해결 -_-;

+++ twe/reset.php
@@ -3,7 +3,18 @@ include "lib.php";
 include "func.php";
 include "schema.php";
 
+$userID = getUserID();
+$userLevel = toInt(getRootDB()->queryFirstField('select userlevel from member where no=%i',$userID));
+
+if($userLevel < 5){
+    echo '관리자 아님';
+    die();
+}
+
 $connect = dbConn();
+
+
+
 
 // 삭제
 unlink("d_setting/set.php");
